Add test feature (#20)
Runs pester tests that reside with the docker image being constructed.

By default, only the tests that reside at the top directory will be run.
This is to account for the scenario where you are traversing a directory structure of repositories that all build using this module. In this scenario we would risk executing tests for docker images below the one we we currently building if we do not limit the recursion when we find the files that pester will execute.