fix #112 - cache docker os type info in a module-internal hashtable
The hashtable approach lets us cache more information in the future.
Because hashtables are also always passed by reference in PowerShell,
they avoid accidental scope bugs where a function that's been passed the
cache *thinks* it's updating it but is really only updating it's
local-scope variable. Hashtables can be passed to functions without this
concern.